minor changes, dated about 9/16

+++ Libraries/ShuffleGameLibrary/CardGame/GameCardGame.cs
@@ -21,7 +21,7 @@ namespace global
             Deck = new CardGamePile("deck");
             for (var i = 0; i < NumberOfCards; i++)
             {
-                Deck.Cards.Add(new CardGameCard(i%13, Math.Floor(i/13)));
+                Deck.Cards.Add(new CardGameCard(i%13, (int) Math.Floor(i/13)));
             }
             for (var i = 0; i < NumberOfJokers; i++)
             {
